merge

+++ kafka-connect/src/main/scala/com/pharbers/kafka/connect/oss/kafka/ConsumerBuilder.scala
@@ -20,13 +20,15 @@ import scala.reflect.ClassTag
   * @author dcs
   * @version 0.0
   * @since 2019/09/26 11:06
-  * @note 一些值得注意的地方
+  * @note 如果kafka非常多，在这儿一直积压，可能会造成oom
+  *       如果消费和生产都特别快，这儿会写本地io，可能造成性能瓶颈
+  *       变成分布式connector后，这儿消费是基于消费组的，要是kafka分片小于worker数量，就会有worker空转
   */
 class ConsumerBuilder[K, V <: SpecificRecordBase](topic: String, classTag: Class[V]) extends Runnable{
     val path = new File("/usr/share/logs-path/msgList")
     val mapping = new ObjectMapper()
     val consumer: KafkaConsumer[K, V] =  new PharbersKafkaConsumer[K, V](List(topic)).getConsumer
-    var keySet: Set[K] = Set()
+    var keys: List[K] = List()
     val msgList = new util.LinkedList[V]()
     if(path.exists()){
         val scanner = new Scanner(path)
@@ -43,37 +45,43 @@ class ConsumerBuilder[K, V <: SpecificRecordBase](topic: String, classTag: Class
         try {
             while(!Thread.currentThread().isInterrupted){
                 consumer.poll(Duration.ofSeconds(1)).asScala.foreach(x => {
-                    if (!keySet.contains(x.key())) {
-                        keySet += x.key()
-                        msgList.add(x.value())
+                    if (!keys.contains(x.key())) {
+                        keys = keys.takeRight(99)
+                        keys :+ x.key()
+                        this.synchronized{
+                            msgList.add(x.value())
+                        }
                     }
                 })
             }
         } catch {
-            case e: InterruptException =>
+            case _: InterruptException =>
             case e: Exception => e.printStackTrace()
-        } finally {
-            consumer.close()
         }
     }
 
     def hasNext: Boolean ={
+        if(msgList.isEmpty) path.delete()
         !msgList.isEmpty
     }
 
     def next: V = {
-        path.delete()
-        path.createNewFile()
-        val write = new FileWriter(path)
-        msgList.asScala.foreach(x => write.write(x.toString + "\n"))
-        write.flush()
-        write.close()
-        keySet = Set()
+        saveMsg()
         msgList.pollFirst()
     }
 
-    def close: Unit ={
+    def saveMsg(): Unit ={
+        path.delete()
+        path.createNewFile()
+        val write = new FileWriter(path)
+        this.synchronized{
+            msgList.asScala.foreach(x => write.write(x.toString + "\n"))
+        }
+        write.flush()
+        write.close()
+    }
+
+    def close(): Unit ={
         consumer.close()
-        next
     }
 }
